Renamed `backend.rs` to `back_end.rs`
Closes https://github.com/PistonDevelopers/glium_graphics/issues/45
