+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "glium_graphics"
-version = "0.0.6"
+version = "0.0.7"
 authors = [
     "Eduard Bopp <eduard.bopp@aepsil0n.de>",
     "bvssvni <bvssvni@gmail.com>"
